Added a first sample of what should become obfs2's threat model.
doc/THREAT_MODEL is a practical list of attacks that obfs2 counters or
fails to counter.

doc/THREAT_MODEL_PHILOSOPHICAL is a non-practical document of wannabe
academia rambling that might prove useful in the future.

Of course both files are not in the correct dir since they only refer
to obfs2 and not to obfsproxy, but hey!