Avoid embedding struct sockaddr in protocol_params_t.
resolve_address_port now heap-allocates a buffer of exactly the
right size for the sockaddr coming back from evutil_getaddrinfo,
and copies a *pointer* to it into an outparam. This removes the
size limit, the need to include <sys/socket.h> in protocol.h,
and code duplicated in all callers of resolve_address_port.
It might make sense to have resolve_address_port return the entire
evutil_addrinfo structure (eliminating the remaining copy) but
let's not go there just yet.