Fix a problem when receiving test files in tester.py.
The workers of tester.py used to socket.recv() till they got
zero-length data. That's usually caused by "EOF" or by the other side
of the socket closing.

A python string (TEST_FILE) don't cause an "EOF" and obfsproxy doesn't
close its sockets since it's supposed to stream data continuously. So,
how did the integration tests work?
tester.py used to socket.shutdown(socket.SHUT_WR) the upstream part of
the connection right after sending the test file, which resulted in
obfsproxy conn_free'ing the whole 'circuit'; including the downstream
side. That used to return zero-length data to recv() and that's how it
was getting out of the ReadWorker:work().

In the case of handshake-less protocols (like 'dummy') that was okay,
since the data were ultimately transferred to the server's destination
by conn_flush_and_free()s.
The thing is that in protocols like obfs2, the handshake couldn't be
completed since the whole 'circuit' was torn down. And that's why
obfs2 tests did not work.

This patch uses socket.timeout instead of zero-length data to get out
of the recv() loop.

Conflicts:

	src/test/tester.py.in