Eliminate connect_with_retry and SIGINT from the Python tests.
The main obfsproxy binary now closes its stdout after it has bound all its
listening sockets; the test suite waits for this before trying to connect.
obfsproxy also interprets EOF on stdin (if stdin is a socket, pipe, or tty)
as a request to shut down cleanly; this is much easier for the test suite to
generate than SIGINT, particularly on Windows.