Enable SOCKS mode in x_rr and chop, to permit testing with Tor itself.
The destination address requested in the SOCKS protocol message is ignored;
the chop/x_rr client will always just connect to the addresses specified on
its command line.  In the future, the mapping from SOCKS request to address
set will be configurable, but we'll want file-based configuration to do that
cleanly.