Defer connection and circuit deallocation to a stable point.
We were crashing because libevent had internally queued up connection
and/or circuit events and would go ahead and fire them even though the
associated objects had been deallocated.  If we instead disable all
events on to-be-deallocated connections/circuits, wait for all of
the current batch of events to be processed, and _then_ deallocate,
the crash doesn't happen.

As a consequence we can no longer test protocol code from the unit
test harness; this is no great loss as we were only testing the null
protocol that way, and the integration tests cover it adequately.

Also fixes a bug in the integration test suite, where we could fail to
wait for a subprocess.