Break up chop.cc in preparation for adding more control blocks.
All of the header formatting code and the reassembly queue move to their
own file.  Some of the low-level transmission and reception code may follow.
This will make it easier to add new types of control blocks (for retransmit,
in-band connection close, the long-awaited rekeying and handshake logic, etc)
and will also make it possible to unit test the separated code.