Stop half-closing downstream connections.
Up till now, when we were done writing to a downstream socket, we
would use shutdown() to send a TCP FIN immediately, even if there
might be data still to be _read_ from that socket.  This turns out to
be unreliable: on the far side, the FIN may cause libevent to signal
read EOF before we are completely done reading data from the network,
causing block loss.  (This is arguably a bug in libevent, but
shutdown() on socket bufferevents is not officially supported at
present, so we need to fix it on our side.)  Also, we have reason to
believe none of our intended cover protocols normally leave sockets
half-closed for any significant length of time, so this might be a
'tell.'  And it has been observed to confuse middleboxes to the point
where they won't pass our traffic at all.

Instead, on the receive side, take note of a half-closed connection if
the cover protocol provides an in-band indication of that fact
(e.g. 'Connection: close' in HTTP) (i.e. conn_t::expect_close() now
does something) but do not call shutdown() on the transmit side.

This exposes a race condition in connection-to-circuit association
which could cause a spurious fatal assertion, and allows us to
simplify the 'should we open new downstream connections?' logic
substantially.

We still half-close our *upstream* sockets when we're done writing to
them.  That should probably change too, but that may require changes
to tltester.